give it its own file
