+++ common-tools/clas-reco/src/main/java/org/jlab/clas/reco/DecoderEngine.java
@@ -4,14 +4,13 @@ import java.util.Set;
 import java.util.HashSet;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.BlockingQueue;
 import org.jlab.clara.base.ClaraUtil;
 import org.jlab.clara.engine.Engine;
 import org.jlab.clara.engine.EngineData;
 import org.jlab.clara.engine.EngineDataType;
 import org.jlab.clara.engine.EngineStatus;
 import org.jlab.detector.decode.CLASDecoder;
+import org.jlab.detector.decode.DecoderPool;
 import org.jlab.io.evio.EvioDataEvent;
 import org.jlab.io.hipo.HipoDataEvent;
 import org.jlab.jnp.hipo4.data.SchemaFactory;
@@ -22,28 +21,6 @@ import org.json.JSONObject;
  * @author baltzell
  */
 public class DecoderEngine implements Engine {
-
-    public static class DecoderPool {
-        BlockingQueue<CLASDecoder> pool;
-        int constantsShared = 64;
-        public DecoderPool(int size, String variation, String timestamp) {
-        	pool = new ArrayBlockingQueue<>(size);
-        	CLASDecoder d0 = null;
-        	for (int i=0; i<size; i++) {
-                CLASDecoder d;
-                if (i % constantsShared == 0) {
-                    d0 = new CLASDecoder();
-                    if (variation != null) d0.setVariation(variation);
-                    if (timestamp != null) d0.setTimestamp(timestamp);
-                    d = d0;
-                }
-                else d = new CLASDecoder(d0);
-                pool.add(d);
-            }
-        }
-        public CLASDecoder take() throws InterruptedException { return pool.take(); }
-        public void put(CLASDecoder d) throws InterruptedException { pool.put(d); }
-    }
 
     static final int POOL_SIZE = 64;
     static final Set<EngineDataType> ED_TYPES = ClaraUtil.buildDataTypes(
