+++ common-tools/clas-reco/src/main/java/org/jlab/clas/reco/EngineProcessor.java
@@ -17,9 +17,9 @@ import org.jlab.utils.options.OptionParser;
 import org.jlab.clara.engine.EngineData;
 import org.jlab.clara.engine.EngineDataType;
 import java.util.Arrays;
-import org.jlab.clas.reco.DecoderEngine.DecoderPool;
 import org.jlab.coda.jevio.EvioException;
 import org.jlab.detector.decode.CLASDecoder;
+import org.jlab.detector.decode.DecoderPool;
 import org.jlab.io.evio.EvioDataEvent;
 import org.jlab.io.evio.EvioSource;
 import org.jlab.io.hipo.HipoDataEvent;
